gen: Make the SCSI and USB drivers compile with the CONFIG_SYS_64BIT_LBA option
A cherry pick, format string editing was required to compile.

These drivers wouldn't compile when the CONFIG_SYS_64BIT_LBA option is turned
on because the used fixed size data types in their exported functions when they
should have used lbaint_t for the block count parameter. That meant that when
the sizes happened to be the same, when using a 28 bit LBA, the driver would
build, but when it wasn't, a 48 bit LBA, things broke.

This change adjusts the signatures to use the right type and makes small
adjustments in the affected functions.

BUG=chrome-os-partner:8180
TEST=Built for emeraldlake2 with the CONFIG_SYS_64BIT_LBA option enabled and
saw its compile errors go away.

Reviewed-on: https://gerrit.chromium.org/gerrit/18058
Reviewed-on: https://gerrit.chromium.org/gerrit/21679
Tested-by: Vadim Bendebury <vbendeb@chromium.org>

+++ disk/part_dos.c
@@ -119,8 +119,9 @@ static void print_partition_extended (block_dev_desc_t *dev_desc, int ext_part_s
 		return;
 	}
 	if(i==DOS_PBR) {
-		printf ("    1\t\t         0\t%10ld\t%2x\n",
-			dev_desc->lba, buffer[DOS_PBR_MEDIA_TYPE_OFFSET]);
+		printf("    1\t\t         0\t%10lld\t%2x\n",
+		       (uint64_t)dev_desc->lba,
+		       buffer[DOS_PBR_MEDIA_TYPE_OFFSET]);
 		return;
 	}
 	/* Print all primary/logical partitions */
